print-proxy-example/requirements.txt
flask==3.0.0
pycups==2.0.1
waitress==3.0.0
